+++ packages/datepicker/src/react/datepicker.tsx
@@ -1,6 +1,6 @@
 import Field from '@pluralsight/ps-design-system-field'
-import { ValueOf } from '@pluralsight/ps-design-system-util'
-import { DateObj, useDayzed } from 'dayzed'
+import { ValueOf, canUseDOM, RefFor } from '@pluralsight/ps-design-system-util'
+import { useDayzed, DateObj } from 'dayzed'
 import React from 'react'
 
 import { Calendar } from './calendar'
@@ -50,8 +50,31 @@ export const DatePicker: React.FC<DatePickerProps> = ({
     setSlide,
     setSelected
   })
+  const ref = React.useRef<HTMLDivElement | undefined>()
+  React.useEffect(() => {
+    if (!canUseDOM()) return () => {}
+
+    const handleClickOutsideMenu = (evt: MouseEvent) => {
+      if (evt.target instanceof HTMLElement) {
+        // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+        if ((ref.current as HTMLDivElement).contains(evt.target)) return
+        setOpen(false)
+      }
+    }
+    document.addEventListener('click', handleClickOutsideMenu, {
+      capture: true
+    })
+    return () =>
+      document.removeEventListener('click', handleClickOutsideMenu, {
+        capture: true
+      })
+  }, [setOpen])
   return (
-    <div style={{ display: 'inline-block', position: 'relative' }} {...props}>
+    <div
+      style={{ display: 'inline-block', position: 'relative' }}
+      {...props}
+      ref={ref as RefFor<'div'>}
+    >
       <TextInputField
         disabled={disabled}
         error={error}