+++ packages/dialog/src/css/index.ts
@@ -1,5 +1,5 @@
 import {
-  colorsBackgroundUtility,
+  colorsBackgroundUtilityCsv,
   colorsPrimaryAction,
   colorsWhite,
   colorsTextIcon,
@@ -139,11 +139,11 @@ export default {
         colorsPrimaryAction.background
       )}`,
       cursor: 'pointer',
-      background: transparentize(0.55, colorsBackgroundUtility.base),
+      background: `rgba(${colorsBackgroundUtilityCsv},0.45)`,
       color: colorsTextIcon.highOnLight
     },
     '&:not([disabled]):active': {
-      background: transparentize(0.35, colorsBackgroundUtility.base),
+      background: `rgba(${colorsBackgroundUtilityCsv},0.65)`,
       color: colorsTextIcon.highOnLight,
       transform: 'scale(0.98)'
     },