+++ packages/menu/src/css/index.ts
@@ -73,17 +73,17 @@ export default {
       outline: 'none'
     },
     '&:focus': {
-      background: colorsBackgroundUtility[25],
+      background: colorsBackgroundUtility[15],
       color: colorsTextIcon.highOnLight
     },
     '&:hover:not([data-disabled])': {
-      background: colorsBackgroundUtility[25],
+      background: colorsBackgroundUtility[15],
       color: colorsTextIcon.highOnLight
     }
   },
   '.psds-menu__list-item--active': {
     '&:not([data-disabled])': {
-      background: colorsBackgroundUtility[25],
+      background: colorsBackgroundUtility[15],
       color: colorsTextIcon.highOnLight
     }
   },
@@ -97,11 +97,11 @@ export default {
     display: 'inline-block',
     width: '100%',
     color: 'inherit',
-    lineHeight: type.lineHeightExtra,
+    lineHeight: type.lineHeightTight,
     fontWeight: type.fontWeight500,
     cursor: 'pointer',
     border: 'none',
-    padding: `${layout.spacingXXSmall} ${layout.spacingMedium}`,
+    padding: `10px ${layout.spacingMedium}`,
     background: 'none',
     textDecoration: 'none',
     outline: 'none',
@@ -155,18 +155,11 @@ export default {
     flexDirection: 'column',
     maxWidth: 248
   },
-  '.psds-menu__item-with-description__name': {
-    display: 'inline-block',
-    color: colorsTextIcon.highOnLight,
-    fontSize: type.fontSize100,
-    lineHeight: type.lineHeightTight,
-    fontWeight: type.fontWeight600
-  },
   '.psds-menu__item-with-description__description': {
-    display: 'inline-block',
-    color: colorsTextIcon.lowOnLight,
-    fontSize: type.fontSize200,
     lineHeight: '16px',
-    fontWeight: type.fontWeightRegular
+    fontSize: type.fontSize100,
+    color: colorsTextIcon.lowOnLight,
+    fontWeight: type.fontWeightRegular,
+    paddingTop: layout.spacingXXSmall
   }
 }