+++ packages/menu/src/react/ItemWithDescription.tsx
@@ -1,4 +1,4 @@
-import { ForwardRefComponent } from '@pluralsight/ps-design-system-util'
+import { forwardRefWithAs } from '@pluralsight/ps-design-system-util'
 import glamorDefault, * as glamorExports from 'glamor'
 import React from 'react'
 
@@ -11,7 +11,6 @@ const styles = {
   item: () => glamor.css(stylesheet['.psds-menu__item-with-description']),
   wrapper: () =>
     glamor.css(stylesheet['.psds-menu__item-with-description__wrapper']),
-  name: () => glamor.css(stylesheet['.psds-menu__item-with-description__name']),
   description: () =>
     glamor.css(stylesheet[`.psds-menu__item-with-description__description`])
 }
@@ -22,15 +21,18 @@ interface MenuItemWithDescriptionProps extends MenuItemProps {
   description: React.ReactText
 }
 
-export const MenuItemWithDescription = React.forwardRef((props, ref) => {
+export const MenuItemWithDescription = forwardRefWithAs<
+  MenuItemWithDescriptionProps,
+  'button'
+>((props, ref) => {
   const { id, name, description, ...rest } = props
   return (
     <Menu.Item {...rest} ref={ref} value={{ id, name }} {...styles.item()}>
       <div {...styles.wrapper()}>
-        <span {...styles.name()}>{name}</span>
+        <span>{name}</span>
         <span {...styles.description()}>{description}</span>
       </div>
       <Menu.Check style={{ marginLeft: 'auto' }} />
     </Menu.Item>
   )
-}) as ForwardRefComponent<'button', MenuItemWithDescriptionProps>
+})