+++ packages/navuser/src/css/index.ts
@@ -1,10 +1,9 @@
 import {
-  colorsBackgroundUtility,
+  colorsBackgroundUtilityCsv,
   colorsTextIcon,
   layout,
   type
 } from '@pluralsight/ps-design-system-core'
-import { transparentize } from '@pluralsight/ps-design-system-util'
 
 export default {
   '.psds-navuser': {
@@ -20,13 +19,13 @@ export default {
     cursor: 'pointer',
 
     '&:hover, &:focus': {
-      background: transparentize(0.85, colorsBackgroundUtility.base)
+      background: `rgba(${colorsBackgroundUtilityCsv},0.15)`
     },
     '&:focus': {
       outline: 'none'
     },
     '&:active': {
-      background: transparentize(0.75, colorsBackgroundUtility.base)
+      background: `rgba(${colorsBackgroundUtilityCsv},0.25)`
     }
   },
   '.psds-navuser__plan-name': {