+++ packages/position/src/react/__stories__/index.story.tsx
@@ -62,8 +62,8 @@ const Box = React.forwardRef<HTMLDivElement, HTMLPropsFor<'div'>>(
       top: '200px',
       left: '200px',
       alignItems: 'center',
-      border: `4px dashed ${core.colorsPink.base}`,
-      color: core.colorsPink.base,
+      border: `4px dashed ${core.colorsPink[6]}`,
+      color: core.colorsPink[6],
       display: 'flex',
       fontSize: core.type.fontSize400,
       fontWeight: core.type.fontWeight700,
@@ -98,7 +98,7 @@ const MockToolip = React.forwardRef<
 
 const ScrollContainer: React.FC = props => {
   const containerSelectors = glamor.css({
-    border: `4px dashed ${core.colorsOrange.base}`,
+    border: `4px dashed ${core.colorsOrange[6]}`,
     color: core.colorsTextIcon.highOnDark,
     height: 500,
     overflow: 'scroll',
@@ -136,9 +136,11 @@ Object.values(positionComponents).forEach(Comp => {
 storiesOf('Components | Position / custom style', module).add(
   'shown element keeps style prop',
   () => (
-    <Above show={<MockToolip style={{ color: core.colorsPink.base }} />}>
+    <positionComponents.Above
+      show={<MockToolip style={{ color: core.colorsPink[6] }} />}
+    >
       <Box>Tooltip is pink</Box>
-    </Above>
+    </positionComponents.Above>
   )
 )
 
@@ -152,13 +154,11 @@ storiesOf('Components | Position / custom ref', module).add(
 
     return (
       <>
-        <Above
-          show={
-            <MockToolip style={{ color: core.colorsPink.base }} ref={ref} />
-          }
+        <positionComponents.Above
+          show={<MockToolip style={{ color: core.colorsPink[6] }} ref={ref} />}
         >
           <Box>Tooltip is pink</Box>
-        </Above>
+        </positionComponents.Above>
         <span style={{ color: 'white', position: 'absolute', top: 250 }}>
           X: {x}, Y: {y}
         </span>
@@ -177,7 +177,7 @@ Object.values(positionComponents).forEach(Comp => {
 
   const Outer: React.FC = props => {
     const selectors = glamor.css({
-      border: `4px dashed ${core.colorsOrange.base}`,
+      border: `4px dashed ${core.colorsOrange[6]}`,
       color: core.colorsTextIcon.highOnDark,
       height: 500,
       padding: 20,