+++ packages/textinput/src/react/index.tsx
@@ -6,8 +6,8 @@ import {
 } from '@pluralsight/ps-design-system-theme'
 import {
   ValueOf,
-  RefForwardingComponent,
-  HTMLPropsFor
+  HTMLPropsFor,
+  forwardRefWithAsAndStatics
 } from '@pluralsight/ps-design-system-util'
 import glamorDefault, * as glamorExports from 'glamor'
 import React from 'react'
@@ -126,14 +126,11 @@ export interface TextInputProps extends Omit<HTMLPropsFor<'input'>, 'size'> {
   value?: React.ReactText
 }
 
-export interface TextInputComponent
-  extends RefForwardingComponent<
-    TextInputProps,
-    HTMLInputElement,
-    TextInputStatics
-  > {}
-
-const TextInput = React.forwardRef(
+const TextInput = forwardRefWithAsAndStatics<
+  TextInputProps,
+  'input',
+  TextInputStatics
+>(
   (
     {
       appearance = vars.appearances.default,
@@ -199,7 +196,7 @@ const TextInput = React.forwardRef(
       </label>
     )
   }
-) as TextInputComponent
+)
 
 TextInput.appearances = vars.appearances
 TextInput.iconAligns = vars.iconAligns