+++ packages/carousel/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.2.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-carousel@11.2.7...@pluralsight/ps-design-system-carousel@11.2.8) (2021-04-05)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-carousel
+
+
+
+
+
+## [11.2.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-carousel@11.2.6...@pluralsight/ps-design-system-carousel@11.2.7) (2021-03-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-carousel
+
+
+
+
+
 ## [11.2.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-carousel@11.2.5...@pluralsight/ps-design-system-carousel@11.2.6) (2021-03-23)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-carousel