+++ packages/datawell/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datawell@5.1.7...@pluralsight/ps-design-system-datawell@5.1.8) (2021-04-05)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-datawell
+
+
+
+
+
+## [5.1.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datawell@5.1.6...@pluralsight/ps-design-system-datawell@5.1.7) (2021-03-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-datawell
+
+
+
+
+
 ## [5.1.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datawell@5.1.5...@pluralsight/ps-design-system-datawell@5.1.6) (2021-03-23)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-datawell