+++ packages/navbar/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navbar@3.0.7...@pluralsight/ps-design-system-navbar@4.0.0) (2021-04-05)
+
+
+### flexibility
+
+* remove base from color scale & add parseToRgb ([#1696](https://github.com/pluralsight/design-system/issues/1696)) ([b5abe62](https://github.com/pluralsight/design-system/commit/b5abe62c164f47cde093b6fce1381af2cb47e21e))
+
+
+### BREAKING CHANGES
+
+* update color.base usages to color[6]
+
+* feat(core): remove colorsBackgroundUtility.base update parseToRgb()
+* switch colorsBackgroundUtility.base usage to
+
+* feat(core): adds colorsBackgroundUtilityCsv
+
+* test(core): update snapshot
+
+
+
+
+
+## [3.0.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navbar@3.0.6...@pluralsight/ps-design-system-navbar@3.0.7) (2021-03-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navbar
+
+
+
+
+
 ## [3.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navbar@3.0.5...@pluralsight/ps-design-system-navbar@3.0.6) (2021-03-23)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-navbar